@@ -3,30 +3,32 @@ description: MrDev ,Inc.
 
 publish_to: 'none'
 
-version: 1.0.0+1
+version: 1.1.0+1
 
 environment:
-  sdk: '>=2.19.4 <3.0.0'
-
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
+  get: ^4.6.6
+  google_fonts: ^6.0.0
+  url_launcher: ^6.2.5
+  lottie: ^2.7.0
+  cupertino_icons: ^1.0.6
+  font_awesome_flutter: ^10.6.0
 
-  get: ^4.6.5
-  google_fonts:
-  url_launcher: ^6.1.10
-
-  lottie: any
-  cupertino_icons: ^1.0.2
-  font_awesome_flutter: any
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  flutter_lints: ^3.0.0
+  flutter_launcher_icons: ^0.13.1
 
-
-  flutter_lints: ^2.0.0
+flutter_launcher_icons:
+  web:
+    generate: true
+    image_path: "assets/images/img.png"
 
 
 flutter:
@@ -34,4 +36,4 @@ flutter:
 
   uses-material-design: true
   assets:
-   - images/
+    - assets/images/
