@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:palstore_welcome_page/app/widgets/build_text.dart';
 
 class LaunchScreen extends StatelessWidget {
-  const LaunchScreen({Key? key}) : super(key: key);
+  const LaunchScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
+    return const Scaffold(
       body: Center(
         child: Column(
           children: [
@@ -22,14 +22,14 @@ class LaunchScreen extends StatelessWidget {
 class _LangButton extends StatelessWidget {
   final String title;
 
-  const _LangButton(this.title, {Key? key}) : super(key: key);
+  const _LangButton(this.title);
 
   @override
   Widget build(BuildContext context) {
     return OutlinedButton(
         onPressed: () {
           //TODO:update app language then go to download page [main screen]
-          final localeKey = title == 'English' ? 'en' : 'ar';
+          // final localeKey = title == 'English' ? 'en' : 'ar';
         },
         child: BuildText(data: title));
   }
