@@ -5,7 +5,7 @@ import '/core/localeSettings/general_local_settings.dart';
 import 'welcome_screen_body.dart';
 
 class WelcomeScreen extends StatelessWidget {
-  const WelcomeScreen({Key? key}) : super(key: key);
+  const WelcomeScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
