@@ -5,21 +5,24 @@ import 'package:palstore_welcome_page/core/constants/store_links.dart';
 import 'package:palstore_welcome_page/core/utils/launcher_utils.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-const _authorText = ' 2023 \u00a9 Author MrDev ,Inc.';
+const _authorText = ' 2025 \u00a9 Author MrDev ,Inc.';
 
 class Footer extends StatelessWidget {
-  const Footer({Key? key}) : super(key: key);
+  const Footer({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return Row(
-      textDirection: TextDirection.ltr,
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        Text(_authorText,
-            style: GoogleFonts.nunito(fontSize: 12.0, color: Colors.white)),
-        _socialRow,
-      ],
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 20.0),
+      child: Row(
+        textDirection: TextDirection.ltr,
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          Text(_authorText,
+              style: GoogleFonts.nunito(fontSize: 12.0, color: Colors.white)),
+          _socialRow,
+        ],
+      ),
     );
   }
 
